+++ docs/changlog/changelog.md
@@ -9,6 +9,9 @@
 
 ### Added
  - Affiliate method to get all parent Affiliates
+ - Auth class
+ - Scope property for Client and Affiliate
+ - Parent Scope property for Affiliate
 
 
 ## [1.0.0] 1/21/2019