+++ docs/changlog/changelog.md
@@ -1,7 +1,12 @@
-## [Unreleased]
+## [1.0.1]
+
+### Updated
+ - Development core url
+
+## [1.0.0]
 
 ### Added
- - Lazy load list and iterator (Not 100% complete but okay for internal use)
+ - Lazy load list and iterator
  - Lazy load iterator which only stores the current page of objects
  - credentials file for a users set of keys
  - Example showing how to use the sdk for quoting items, creating an order, and charging the order.